refactor: move code in one file App.tsx
