--- src/components/header.css
@@ -1,11 +1,0 @@
-.header {
-  padding: 2em 0;
-  background-color: #ffebcd;
-}
-
-.header__title {
-  margin: 0;
-  padding-bottom: 1em;
-  font-size: 3em;
-  text-align: center;
-}
