--- src/components/search-form.css
@@ -1,7 +1,0 @@
-.search-form {
-  text-align: center;
-}
-
-.search-form__label {
-  font-size: 1.6em;
-}
